+++ NEWS.md
@@ -17,6 +17,7 @@
 ## New supported models
 
 - Added support for `vgam` (*VGAM*), `complmrob` (*complmrob*), `cpglm`, `cpglmm` (*cplm*), `fixest` (*fixest*), `feglm` (*alpaca*), `glmx` (*glmx*), `glmmadmb` (*glmmADMB*), `mcmc` (*coda*), `mixor` (*mixor*).
+- `model_parameters()` now supports `blavaan` models (*blavaan*).
 
 ## General